add modification

+++ src/main/java/hei/arit/demo/Services/AddNumbersService.java
@@ -4,8 +4,8 @@ import org.springframework.stereotype.Service;
 
 @Service
 public class AddNumbersService {
-  public int sumNumbers(int a, int b) {
-    int result = a + b;
+  public long sumNumbers(long a, long b) {
+    long result = a + b;
     if (result < 0) {
       throw new IllegalArgumentException("sum of a and b is invalide");
     }
